Don't do develop builds for now

+++ build_scripts/update_cvmfs_artdaq.sh
@@ -5,6 +5,7 @@ git config --global --add safe.directory '*'
 artVer=${artVer:-s132}
 artdaqVer=${artdaqVer:-v3_16_00}
 force=${force:-0}
+develop=${develop:-0}
 
 function cleanup() {
     (
@@ -26,14 +27,16 @@ if [ $force -eq 1 ] || ! [ -d /cvmfs/fermilab.opensciencegrid.org/products/artda
     cleanup
 fi
 
-echo "Building artdaq-develop"
-cd /cvmfs/fermilab.opensciencegrid.org/products/artdaq/spack_areas
-mkdir artdaq-develop;cd artdaq-develop
-touch .cvmfscatalog
-rm quick-spack-start.sh*;wget https://raw.githubusercontent.com/art-daq/artdaq_demo/refs/heads/develop/tools/quick-spack-start.sh && chmod +x quick-spack-start.sh
-./quick-spack-start.sh --develop --no-kmod --arch linux-almalinux9-x86_64_v3 \
-                       --upstream /cvmfs/fermilab.opensciencegrid.org/products/artdaq/spack_areas/artdaq-$artdaqVer \
-                       --upstream /cvmfs/fermilab.opensciencegrid.org/products/artdaq/spack_areas/art-suite-$artVer
-cleanup
+if [ $develop -eq 1 ];then
+    echo "Building artdaq-develop"
+    cd /cvmfs/fermilab.opensciencegrid.org/products/artdaq/spack_areas
+    mkdir artdaq-develop;cd artdaq-develop
+    touch .cvmfscatalog
+    rm quick-spack-start.sh*;wget https://raw.githubusercontent.com/art-daq/artdaq_demo/refs/heads/develop/tools/quick-spack-start.sh && chmod +x quick-spack-start.sh
+    ./quick-spack-start.sh --develop --no-kmod --arch linux-almalinux9-x86_64_v3 \
+                           --upstream /cvmfs/fermilab.opensciencegrid.org/products/artdaq/spack_areas/artdaq-$artdaqVer \
+                           --upstream /cvmfs/fermilab.opensciencegrid.org/products/artdaq/spack_areas/art-suite-$artVer
+    cleanup
+fi
 
 git config --global --unset-all safe.directory
